requirements-optional.txt
# Optional cloud storage providers
boto3>=1.34.0
google-cloud-storage>=2.16.0
azure-storage-blob>=12.20.0
